+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - improve date handling in `apt.inspector` @alihajqani
 - add `MCP_CLIENT_TEST_TOKEN` to `.env.example` @alihajqani
 - read `Authorization` in `mcp-client` from `.env` @alihajqani
+- feat(`mcp-client`): enhance resource handling and add resource reading functionality @alihajqani
 
 ### Changed